Fix handling of smpc and destination for next element when multiple elements are in data_outgoing
The tuple appended to the data_outgoing list in AppState.send_data_to_participant, AppState.send_data_to_coordinator and AppState.broadcast_data has the structure
(data, smpc, destination). This is not reflected in the parsing of the tuple in App.handle_outgoing, where the smpc field is read into the destination attr and
the destination field is read into the smpc attr. This commit fixes this behaviour in App.handle_outgoing.