Fix behavior when running with only one client
When running with only one client for testing purposes, gather_data behaves differently as it gives the unwrapped data. Specific app code should not be forced to handle this.
The await_data method should keep its behavior in the case that only data from one client is expected as when listening to a coordinator broadcast. For that a default value is set.